Workshop's program (#339)
* add custom message source

* Add workshop example with send/reply contracts

* yaml test

* rename test

* refactor

* add value check in gtest

* rename contract
use async

* fix

* parse json from init payload

* Add code, rm State methods

* add admins, set config from handle

* use cloned state to prevent refcell from unnecessary borrow

* use metadata

* metadata

* refactor

* Update examples/token-vendor/contract-example-reply/src/lib.rs


* convert ss58 addr to ProgramId

* commit a state change only after sending a message

* Remove previous token-vendor

* Rewrite contract

* Provide contract example

* Update js lib

* Update encode script

* Vendor v0.1.0

* Fmt

* Logic fixes

* Fix script

* Change paths for gtest workshop

* Fixes

* Remove unused mutex

* Fix

+++ gtest/src/sample.rs
@@ -53,7 +53,7 @@ pub struct Program {
     pub source: Option<SourceVariant>,
     pub init_message: Option<PayloadVariant>,
     pub init_gas_limit: Option<u64>,
-    pub init_value: Option<u64>,
+    pub init_value: Option<u128>,
 }
 
 #[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
@@ -193,7 +193,7 @@ pub struct Message {
     pub init: Option<bool>,
     pub payload: Option<PayloadVariant>,
     pub gas_limit: Option<u64>,
-    pub value: Option<u64>,
+    pub value: Option<u128>,
     #[serde(rename = "exitCode")]
     pub exit_code: Option<i32>,
 }
